backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0
pytest==7.3.1
gunicorn==20.1.0
# For notifications in Phase 2
# twilio==8.1.0
